requirements.txt
PyQt5
pytube
ffmpeg-python